Abstract out WorkflowInputBox
