+++ src/pages/HomePage.js
@@ -1,8 +1,13 @@
 import React from 'react'
-import { Box } from 'grommet'
+import { Anchor, Box, Heading, Paragraph } from 'grommet'
 import { observer } from 'mobx-react'
 import { withRouter } from 'react-router-dom'
+import styled from 'styled-components'
 import AppContext from 'stores'
+
+const ConstrainedParagraph = styled(Paragraph)`
+  max-width: 40rem;
+`
 
 function HomePage () {
   const store = React.useContext(AppContext)
@@ -13,7 +18,8 @@ function HomePage () {
   
   return (
     <Box>
-      <h1>Home</h1>
+      <Heading as="h2" size="xsmall">Welcome to Zoo Notes!</Heading>
+      <ConstrainedParagraph>...</ConstrainedParagraph>
     </Box>
   )
 }
